+++ src/main/java/io/backend/project0/service/ObjectStoredService.java
@@ -1,20 +1,20 @@
 package io.backend.project0.service;
 
 import io.backend.project0.StorageDir;
-import io.backend.project0.entity.Bucket;
-import io.backend.project0.entity.ObjectStored;
 import io.backend.project0.entity.ObjectPart;
+import io.backend.project0.entity.ObjectStored;
 import io.backend.project0.repository.BucketRepository;
-import io.backend.project0.repository.ObjectStoredRepository;
 import io.backend.project0.repository.ObjectPartRepository;
-import org.apache.commons.codec.DecoderException;
-import org.apache.commons.codec.binary.Hex;
+import io.backend.project0.repository.ObjectStoredRepository;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.io.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.io.*;
+import java.io.ByteArrayOutputStream;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
@@ -40,6 +40,8 @@ public class ObjectStoredService {
         if (isObjectExist(name, bucketName)) return null;
         long createdTime = System.currentTimeMillis();
         ObjectStored objectStored = new ObjectStored(name,createdTime,createdTime,bucketName);
+        //Recheck spamming
+        if (isObjectExist(name, bucketName)) return null;
         objectStoredRepository.save(objectStored);
 
         return objectStored;
@@ -51,12 +53,12 @@ public class ObjectStoredService {
 
     public ObjectStored complete(String name, String bucketName){
         ObjectStored objectStored = objectStoredRepository.findByObjectNameAndBucketName(name,bucketName);
-        long size =0;
 
-        if(!objectStored.isComplete()) {
+        if(objectStored!=null && !objectStored.isComplete()) {
             List<ObjectPart> objectParts = objectPartRepository.findAllByBucketNameAndObjectName(bucketName, name);
             objectParts.sort(Comparator.comparingInt(ObjectPart::getPartNumber));
             ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+            long size =0;
 
             try {
                 for (ObjectPart p : objectParts) {
@@ -72,6 +74,7 @@ public class ObjectStoredService {
                 objectStored.setSize(size);
                 objectStored.setModified(curTime);
                 objectStored.setComplete(true);
+                objectStored.setPath(StorageDir.storage+'/'+bucketName+'/'+name);
                 objectStoredRepository.save(objectStored);
                 return objectStored;
 
@@ -85,8 +88,10 @@ public class ObjectStoredService {
 
     public void deleteObject(String objectName, String bucketName){
         ObjectStored objectStored = objectStoredRepository.findByObjectNameAndBucketName(objectName,bucketName);
-        objectStoredRepository.delete(objectStored);
-        objectPartRepository.deleteAllByBucketNameAndObjectName(bucketName,objectName);
+        if (objectStored!=null) {
+            objectStoredRepository.delete(objectStored);
+            objectPartRepository.deleteAllByBucketNameAndObjectName(bucketName, objectName);
+        }
     }
 
     public ObjectStored getObject(String objectName, String bucketName){
@@ -95,28 +100,35 @@ public class ObjectStoredService {
 
     public void addAndUpdateMetadata(String objectName, String bucketName, String key, String value){
         ObjectStored objectStored = objectStoredRepository.findByObjectNameAndBucketName(objectName,bucketName);
-        HashMap<String,String> metadata = objectStored.getMetadata();
-        metadata.put(key,value);
-        objectStored.setMetadata(metadata);
-        objectStoredRepository.save(objectStored);
+        if (objectStored!=null) {
+            HashMap<String, String> metadata = objectStored.getMetadata();
+            metadata.put(key, value);
+            objectStored.setMetadata(metadata);
+            objectStoredRepository.save(objectStored);
+        }
     }
 
     public void deleteMetadata(String objectName, String bucketName,String key){
         ObjectStored objectStored = objectStoredRepository.findByObjectNameAndBucketName(objectName,bucketName);
-        HashMap<String,String> metadata = objectStored.getMetadata();
-        metadata.remove(key);
-        objectStored.setMetadata(metadata);
-        objectStoredRepository.save(objectStored);
+        if (objectStored!=null) {
+            HashMap<String, String> metadata = objectStored.getMetadata();
+            metadata.remove(key);
+            objectStored.setMetadata(metadata);
+            objectStoredRepository.save(objectStored);
+        }
     }
 
     public String getMetadata(String objectName, String bucketName,String key){
-        HashMap<String,String > metadata = objectStoredRepository.findByObjectNameAndBucketName(objectName,bucketName).getMetadata();
-        if (metadata.containsKey(key))return metadata.get(key);
+        ObjectStored objectStored = objectStoredRepository.findByObjectNameAndBucketName(objectName,bucketName);
+        if (objectStored !=null) return objectStored.getMetadata().get(key);
         return "";
     }
 
     public HashMap<String,String> getAllMetadata(String objectName, String bucketName){
-        return objectStoredRepository.findByObjectNameAndBucketName(objectName,bucketName).getMetadata();
+        HashMap<String, String> metadata = new HashMap<>();
+        ObjectStored objectStored = objectStoredRepository.findByObjectNameAndBucketName(objectName,bucketName);
+        if (objectStored!=null) metadata = objectStored.getMetadata();
+        return metadata;
     }
 
     public boolean validateObjectName(String objectName){
